fixed syntax errors
